requirements.txt
async-timeout==4.0.3
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.11.0
docopt==0.6.2
filelock==3.12.4
Flask==2.3.3
Flask-Session==0.5.0
flatbuffers==23.5.26
fonttools==4.42.1
gunicorn==21.2.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
onnx==1.14.1
onnxruntime==1.15.1
Pillow==10.0.0
protobuf==4.24.3
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
redis==5.0.0
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.0.0
torchvision==0.15.1
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
yarg==0.1.9

